+++ src/components/Function/index.jsx
@@ -1,16 +1,14 @@
 import React from 'react'
-//函数中没有componentDidMount
-const Demo = () => {
-    const [count,setCount]=React.useState(0)
-    React.useEffect(()=>{
-        setInterval(()=>{
-            setCount(count=>count+1)
-        },1000)
-    },[])//[]代表谁也不检测，可以监测count，变化是刷新，数字会越来越快
+
+const Demo= () => {
+    const myRef=React.useRef()
+    const show=()=>{
+        alert(myRef.current.value)
+    }
     return (
         <div>
-            <h2>The current value is :{count}</h2>
-            
+            <input type="text" ref={myRef}/>
+            <button onClick={show}>show the data</button>
         </div>
     )
 }
